Changing co-ordinate capture to int instead fo float for mouse events.

+++ Engine/Src/Core/Event/Mouse/MouseScrolledEvent.h
@@ -7,11 +7,11 @@ namespace Vkr
     class MouseScrolledEvent : public Event
     {
     public:
-        MouseScrolledEvent(const bool direction, const f32 xOffset, const f32 yOffset)
+        MouseScrolledEvent(const bool direction, const i32 xOffset, const i32 yOffset)
             : direction(direction), mXOffset(xOffset), mYOffset(yOffset) {}
 
-        inline f32 GetXOffset() const { return mXOffset; }
-        inline f32 GetYOffset() const { return mYOffset; }
+        inline i32 GetXOffset() const { return mXOffset; }
+        inline i32 GetYOffset() const { return mYOffset; }
         inline bool GetDirection() const { return direction; }
 
         virtual int GetCategoryFlags() const override
@@ -24,6 +24,6 @@ namespace Vkr
     private:
         // `true` -> mouse scrolled up else mouse scrolled down.
         const bool direction;
-        const f32 mXOffset, mYOffset;
+        const i32 mXOffset, mYOffset;
     };
 }
